[refactor] #18 Project와 ProjectUser 패키지 분리 및 주석 삭제 (#50)
* [refactor] 코드 설명 주석 삭제

* [refactor] Project와 ProjectUser 패키지 분리
