+++ module-domain/src/main/java/com/welcommu/moduledomain/projectUser/ProjectUserManageRole.java
@@ -1,4 +1,4 @@
-package com.welcommu.moduledomain.project;
+package com.welcommu.moduledomain.projectUser;
 
 public enum ProjectUserManageRole {
     CLIENT_USER,
