+++ module-repository/src/main/java/com/welcommu/modulerepository/projectpost/repository/ProjectPostRepository.java
@@ -1,6 +1,6 @@
 package com.welcommu.modulerepository.projectpost.repository;
 
-import com.welcommu.moduledomain.projectpost.entity.ProjectPost;
+import com.welcommu.moduledomain.projectpost.ProjectPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
