+++ module-service/src/test/java/com/welcommu/moduleservice/project/ProjectServiceTest.java
@@ -1,6 +1,7 @@
 package com.welcommu.moduleservice.project;
 
 import com.welcommu.moduledomain.project.*;
+import com.welcommu.moduledomain.projectUser.ProjectUser;
 import com.welcommu.moduledomain.user.User;
 import com.welcommu.modulerepository.company.CompanyRepository;
 import com.welcommu.modulerepository.project.ProjectRepository;
